combines ustatus
